workflow/requirements.txt
temporalio==1.5.0
psycopg2-binary==2.9.9
pyyaml==6.0.1
langchain==0.3.19
langchain-community==0.3.18
langchain-openai==0.3.7
openai==1.64.0